+++ pages/auth/login.tsx
@@ -1,15 +1,26 @@
-import React from "react";
+import React, { useState } from "react";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import SignInKey from "@components/login/SignInKey";
-import { useTranslation } from "next-i18next";
+import TemporaryToken from "@components/login/TemporaryToken";
 import { GetServerSideProps } from "next";
 
+export interface LoginStageProps {
+  setParentStage: (stage: number) => void;
+}
+
 const Login = () => {
-  const { t } = useTranslation("login");
+  const [loginState, setLoginState] = useState(1);
+
+  const setStage = (stage: number) => {
+    setLoginState(stage);
+  };
 
   return (
     <>
-      <SignInKey />
+      <div>
+        {loginState === 1 && <SignInKey setParentStage={setStage} />}
+        {loginState === 2 && <TemporaryToken setParentStage={setStage} />}
+      </div>
     </>
   );
 };